requirements.txt
streamlit
ultralytics
pymongo
bson
pillow
numpy
opencv-python-headless
pandas
python-dotenv